requirements.txt
flask>=3.0
qrcode[pil]>=7.4
Pillow>=10.0
reportlab>=4.0
